requirements.txt
streamlit>=1.28.0
langchain>=0.0.300
langgraph>=0.0.23
llama-index>=0.9.0
openai>=1.3.0
google-cloud-documentai>=2.24.0
google-cloud-storage>=2.13.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.2
wikipedia>=1.4.0
scholarly>=1.7.11
pandas>=2.1.1
numpy>=1.26.0
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0
pillow>=10.1.0
pydantic>=2.4.2
tiktoken>=0.5.1
pypdf>=3.17.0
